removed unnecessary files and updated css and components
