+++ client/src/App.css
@@ -85,7 +85,7 @@ button {
   padding: 20px;
 }
 
-.main-box, .templates-box, .holiday-box {
+.main-box, .templates-box {
   display: grid;
   gap: 50px; 
   grid-template-columns: repeat(1, 1fr);
@@ -93,6 +93,16 @@ button {
   padding: 20px;
   padding-bottom: 100px;
   max-width: 1000px;
+}
+
+.holiday-box {
+  display: grid;
+  gap: 50px; 
+  grid-template-columns: repeat(1, 1fr);
+  margin-inline: auto;
+  padding: 20px;
+  padding-bottom: 100px;
+  max-width: auto;
 }
 
 .bespoke-box-button {
@@ -411,6 +421,26 @@ label {
   align-items: center;
   justify-content: center;
   margin: 20px;
+  margin-left: 0px;
+  margin-right: 0px;
   border-radius: 20px;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  box-shadow: 0px 4px 4px rgba(109, 15, 64, 0.25);
+}
+
+.holiday-buttons {
+  display: flex;
+  flex-direction: row;
+  align-content: space-between;
+  justify-content: space-between;
+  margin: 20px;
+}
+
+.editButton, .deleteButton {
+  border: 1px solid white;
+  border-radius: 6px;
+  padding: 6px 14px;
+  color: white;
+  font-size: 16px;
+  font-weight: 600;
+  background-image: linear-gradient(to right, rgba(247, 73, 168, 0.828), rgb(247, 176, 114));
 }
